requirements.txt
fastapi
uvicorn
requests
python-dotenv
faster-whisper
opencv-python
mediapipe
streamlit
matplotlib
